restore: switch ntrec size to int #159
We use int in the common/dump code already.  Trying to use it as long
in restore creates size confusion between the different modules.

Fixes: https://sourceforge.net/p/dump/bugs/159

+++ restore/main.c
@@ -92,7 +92,7 @@ int	dokerberos = 0;
 char	command = '\0';
 long	dumpnum = 1;
 long	volno = 0;
-long	ntrec;
+int	ntrec;
 char	*dumpmap = NULL;
 char	*usedinomap = NULL;
 dump_ino_t maxino;
